Project Fast API/requirements.txt
fastapi
uvicorn
sqlalchemy
mysql-connector-python
<<<<<<< Updated upstream
pydantic
=======
pydantic[email]
python-dotenv
>>>>>>> Stashed changes
